[JENKINS-31772] Factor out common code to a superclass.
